@@ -49,7 +49,7 @@ class ProcessorCollectionViewController: UICollectionViewController {
         (BlackWhiteProcessor(), "B&W"),
         (CroppingImageProcessor(size: CGSize(width: 100, height: 100)), "Cropping"),
         (DownsamplingImageProcessor(size: CGSize(width: 25, height: 25)), "Downsampling"),
-        (BlurImageProcessor(blurRadius: 5) >> RoundCornerImageProcessor(cornerRadius: 20), "Blur + Round Corner")
+        (BlurImageProcessor(blurRadius: 5) |> RoundCornerImageProcessor(cornerRadius: 20), "Blur + Round Corner")
     ]
     
     override func viewDidLoad() {
