@@ -47,7 +47,10 @@ class InterfaceController: WKInterfaceController {
     
     func refreshImage() {
         let url = URL(string: "https://raw.githubusercontent.com/onevcat/Kingfisher/master/images/kingfisher-\(currentIndex! + 1).jpg")!
-        interfaceImage.kf.setImage(with: url)
+        print("Start loading... \(url)")
+        interfaceImage.kf.setImage(with: url) { r in
+            print(r)
+        }
     }
 
     override func willActivate() {
