@@ -24,6 +24,8 @@
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 //  THE SOFTWARE.
 
+#if !os(watchOS)
+
 import CoreImage
 
 // Reuse the same CI Context for all CI drawing.
@@ -48,12 +50,12 @@ extension CIImageProcessor {
     /// - Returns: The processed image.
     ///
     /// - Note: See documentation of `ImageProcessor` protocol for more.
-    public func process(item: ImageProcessItem, options: KingfisherParsedOptionsInfo) -> Image? {
+    public func process(item: ImageProcessItem, options: KingfisherParsedOptionsInfo) -> KFCrossPlatformImage? {
         switch item {
         case .image(let image):
             return image.kf.apply(filter)
         case .data:
-            return (DefaultImageProcessor.default >> self).process(item: item, options: options)
+            return (DefaultImageProcessor.default |> self).process(item: item, options: options)
         }
     }
 }
@@ -69,7 +71,7 @@ public struct Filter {
     }
     
     /// Tint filter which will apply a tint color to images.
-    public static var tint: (Color) -> Filter = {
+    public static var tint: (KFCrossPlatformColor) -> Filter = {
         color in
         Filter {
             input in
@@ -105,7 +107,7 @@ public struct Filter {
     }
 }
 
-extension KingfisherWrapper where Base: Image {
+extension KingfisherWrapper where Base: KFCrossPlatformImage {
 
     /// Applies a `Filter` containing `CIImage` transformer to `self`.
     ///
@@ -115,7 +117,7 @@ extension KingfisherWrapper where Base: Image {
     /// - Note:
     ///    Only CG-based images are supported. If any error happens
     ///    during transforming, `self` will be returned.
-    public func apply(_ filter: Filter) -> Image {
+    public func apply(_ filter: Filter) -> KFCrossPlatformImage {
         
         guard let cgImage = cgImage else {
             assertionFailure("[Kingfisher] Tint image only works for CG-based image.")
@@ -135,8 +137,10 @@ extension KingfisherWrapper where Base: Image {
         #if os(macOS)
             return fixedForRetinaPixel(cgImage: result, to: size)
         #else
-            return Image(cgImage: result, scale: base.scale, orientation: base.imageOrientation)
+            return KFCrossPlatformImage(cgImage: result, scale: base.scale, orientation: base.imageOrientation)
         #endif
     }
 
 }
+
+#endif
