@@ -24,7 +24,9 @@
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 //  THE SOFTWARE.
 
-#if canImport(AppKit)
+#if !os(watchOS)
+
+#if canImport(AppKit) && !targetEnvironment(macCatalyst)
 import AppKit
 public typealias IndicatorView = NSView
 #else
@@ -118,11 +120,22 @@ final class ActivityIndicator: Indicator {
             activityIndicatorView.controlSize = .small
             activityIndicatorView.style = .spinning
         #else
+            let indicatorStyle: UIActivityIndicatorView.Style
+
             #if os(tvOS)
-                let indicatorStyle = UIActivityIndicatorView.Style.white
+            if #available(tvOS 13.0, *) {
+                indicatorStyle = UIActivityIndicatorView.Style.large
+            } else {
+                indicatorStyle = UIActivityIndicatorView.Style.white
+            }
             #else
-                let indicatorStyle = UIActivityIndicatorView.Style.gray
+            if #available(iOS 13.0, * ) {
+                indicatorStyle = UIActivityIndicatorView.Style.medium
+            } else {
+                indicatorStyle = UIActivityIndicatorView.Style.gray
+            }
             #endif
+
             #if swift(>=4.2)
             activityIndicatorView = UIActivityIndicatorView(style: indicatorStyle)
             #else
@@ -132,10 +145,22 @@ final class ActivityIndicator: Indicator {
     }
 }
 
+#if canImport(UIKit)
+extension UIActivityIndicatorView.Style {
+    #if compiler(>=5.1)
+    #else
+    static let large = UIActivityIndicatorView.Style.white
+    #if !os(tvOS)
+    static let medium = UIActivityIndicatorView.Style.gray
+    #endif
+    #endif
+}
+#endif
+
 // MARK: - ImageIndicator
 // Displays an ImageView. Supports gif
 final class ImageIndicator: Indicator {
-    private let animatedImageIndicatorView: ImageView
+    private let animatedImageIndicatorView: KFCrossPlatformImageView
 
     var view: IndicatorView {
         return animatedImageIndicatorView
@@ -156,7 +181,7 @@ final class ImageIndicator: Indicator {
             return nil
         }
 
-        animatedImageIndicatorView = ImageView()
+        animatedImageIndicatorView = KFCrossPlatformImageView()
         animatedImageIndicatorView.image = image
         
         #if os(macOS)
@@ -186,3 +211,5 @@ final class ImageIndicator: Indicator {
         animatedImageIndicatorView.isHidden = true
     }
 }
+
+#endif
