@@ -29,7 +29,7 @@ import XCTest
 
 class ImageViewExtensionTests: XCTestCase {
 
-    var imageView: ImageView!
+    var imageView: KFCrossPlatformImageView!
     
     override class func setUp() {
         super.setUp()
@@ -44,7 +44,7 @@ class ImageViewExtensionTests: XCTestCase {
     override func setUp() {
         super.setUp()
 
-        imageView = ImageView()
+        imageView = KFCrossPlatformImageView()
         KingfisherManager.shared.downloader = ImageDownloader(name: "testDownloader")
         KingfisherManager.shared.defaultOptions = [.waitForCache]
         
@@ -179,7 +179,7 @@ class ImageViewExtensionTests: XCTestCase {
         }
         
         group.enter()
-        let anotherImageView = ImageView()
+        let anotherImageView = KFCrossPlatformImageView()
         anotherImageView.kf.setImage(with: url) { result in
             XCTAssertNotNil(result.value)
             group.leave()
@@ -237,6 +237,7 @@ class ImageViewExtensionTests: XCTestCase {
         
         cache1.clearDiskCache()
         cache2.clearDiskCache()
+        cleanDefaultCache()
         
         let exp = expectation(description: #function)
         let url = testURLs[0]
@@ -274,7 +275,7 @@ class ImageViewExtensionTests: XCTestCase {
     
     func testCustomizeStructIndicatorExisting() {
         struct StructIndicator: Indicator {
-            let view = View()
+            let view = KFCrossPlatformView()
             func startAnimatingView() {}
             func stopAnimatingView() {}
         }
@@ -313,7 +314,7 @@ class ImageViewExtensionTests: XCTestCase {
         
         imageView.kf.indicatorType = .image(imageData: testImageData)
         XCTAssertTrue(imageView.kf.indicator is ImageIndicator)
-        let image = (imageView.kf.indicator?.view as? ImageView)?.image
+        let image = (imageView.kf.indicator?.view as? KFCrossPlatformImageView)?.image
         XCTAssertNotNil(image)
         XCTAssertTrue(image!.renderEqual(to: testImage))
         
@@ -434,7 +435,7 @@ class ImageViewExtensionTests: XCTestCase {
         XCTAssertEqual(testImage, imageView.image)
         
         // While current image is not nil, keep it
-        let anotherImage = Image(data: testImageJEPGData)
+        let anotherImage = KFCrossPlatformImage(data: testImageJEPGData)
         imageView.image = anotherImage
         imageView.kf.setImage(with: url, placeholder: testImage, options: [.keepCurrentImageWhileLoading]) { result in
             XCTAssertNotEqual(self.imageView.image, anotherImage)
@@ -527,7 +528,7 @@ class ImageViewExtensionTests: XCTestCase {
 
         stub(url, data: testImageData, length: 123)
 
-        let emptyImage = Image()
+        let emptyImage = KFCrossPlatformImage()
         var processBlockCalled = false
 
         imageView.kf.setImage(
@@ -553,7 +554,7 @@ class ImageViewExtensionTests: XCTestCase {
 
         stub(url, data: testImageData, length: 123)
 
-        let view = View()
+        let view = KFCrossPlatformView()
         var processBlockCalled = false
 
         imageView.kf.setImage(
@@ -653,7 +654,7 @@ class ImageViewExtensionTests: XCTestCase {
         waitForExpectations(timeout: 3, handler: nil)
     }
     
-    func testImageCacheExtendingExpirationTask() {
+    func testMemoryImageCacheExtendingExpirationTask() {
         let exp = expectation(description: #function)
         let url = testURLs[0]
         stub(url, data: testImageData)
@@ -688,7 +689,7 @@ class ImageViewExtensionTests: XCTestCase {
         waitForExpectations(timeout: 3, handler: nil)
     }
     
-    func testImageCacheNotExtendingExpirationTask() {
+    func testMemoryImageCacheNotExtendingExpirationTask() {
         let exp = expectation(description: #function)
         let url = testURLs[0]
         stub(url, data: testImageData)
@@ -720,6 +721,73 @@ class ImageViewExtensionTests: XCTestCase {
         
         waitForExpectations(timeout: 3, handler: nil)
     }
+
+    func testDiskImageCacheExtendingExpirationTask() {
+        let exp = expectation(description: #function)
+        let url = testURLs[0]
+        stub(url, data: testImageData)
+
+        let options: KingfisherOptionsInfo = [.memoryCacheExpiration(.expired),
+                                              .diskCacheExpiration(.seconds(2)),
+                                              .diskCacheAccessExtendingExpiration(.expirationTime(.seconds(100)))]
+
+        imageView.kf.setImage(with: url, options: options) { result in
+            XCTAssertNotNil(result.value?.image)
+            XCTAssertTrue(result.value!.cacheType == .none)
+
+            delay(1, block: {
+                self.imageView.kf.setImage(with: url, options: options) { result in
+                    XCTAssertNotNil(result.value?.image)
+                    XCTAssertTrue(result.value!.cacheType == .disk)
+                    delay(2, block: {
+                        self.imageView.kf.setImage(with: url, options: options) { result in
+                            XCTAssertNotNil(result.value?.image)
+                            XCTAssertTrue(result.value!.cacheType == .disk)
+
+                            exp.fulfill()
+                        }
+                    })
+                }
+            })
+        }
+
+        waitForExpectations(timeout: 5, handler: nil)
+    }
+
+    func testDiskImageCacheNotExtendingExpirationTask() {
+        let exp = expectation(description: #function)
+        let url = testURLs[0]
+        stub(url, data: testImageData)
+
+        let options: KingfisherOptionsInfo = [.memoryCacheExpiration(.expired),
+                                              .diskCacheExpiration(.seconds(2)),
+                                              .diskCacheAccessExtendingExpiration(.none)]
+
+        imageView.kf.setImage(with: url, options: options) { result in
+            XCTAssertNotNil(result.value?.image)
+            XCTAssertTrue(result.value!.cacheType == .none)
+
+            delay(1, block: {
+                self.imageView.kf.setImage(with: url, options: options) { result in
+                    XCTAssertNotNil(result.value?.image)
+                    XCTAssertTrue(result.value!.cacheType == .disk)
+
+                        delay(2, block: {
+                            self.imageView.kf.setImage(with: url, options: options) { result in
+                                XCTAssertNotNil(result.value?.image)
+                                XCTAssertTrue(result.value!.cacheType == .none)
+
+                                exp.fulfill()
+                            }
+                        })
+                }
+            })
+        }
+
+        waitForExpectations(timeout: 5, handler: nil)
+    }
+
+
 }
 
-extension View: Placeholder {}
+extension KFCrossPlatformView: Placeholder {}
